Break S3 tests into own spec file
